refactor: Move the usePapersDefinitions hook to its Provider file
