+++ src/components/Papers/Empty/EmptyNoHeader.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { useNavigate, useParams, useLocation } from 'react-router-dom'
 import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
-import { usePapersDefinitions } from 'src/components/Hooks/usePapersDefinitions'
+import { usePapersDefinitions } from 'src/components/Contexts/PapersDefinitionsProvider'
 import { makeCountrySearchParam } from 'src/components/Papers/Empty/helpers'
 import styles from 'src/components/Papers/Empty/styles.styl'
 import HarvestBanner from 'src/components/Papers/HarvestBanner'
